# api/openapi.yaml
openapi: 3.1.0
info:
  title: NovaForge API
  version: 0.1.0
  description: >-
    The NovaForge control plane. The web GUI is deferred, so this contract is
    the complete description of the platform for every client: the nf CLI,
    external agents, and the future GUI.
servers:
  - url: /
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: No usable credentials were presented
    Forbidden:
      description: The caller's organization scope does not permit this
    NotFound:
      description: No such resource in the caller's organization
paths:
...
  /api/v1/orgs:
    get:
      operationId: listOrgs
      summary: Organizations the caller belongs to
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createOrg
      summary: Create an organization
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}:
    delete:
      operationId: deleteOrg
      summary: Delete an organization and everything held for it (owner only, name confirmed)
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      operationId: getOrg
      summary: One organization
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/members:
    get:
      operationId: listOrgMembers
      summary: Organization members
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: addOrgMember
      summary: Add an organization member
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos:
    get:
      operationId: listRepos
      summary: Repositories in an organization
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createRepo
      summary: Create a repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}:
    delete:
      operationId: deleteRepo
      summary: Delete a repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      operationId: getRepo
      summary: One repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/blob/{ref}/{path}:
    get:
      operationId: getBlob
      summary: File contents
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: ref
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/branches:
    get:
      operationId: listBranches
      summary: Branches
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createBranch
      summary: Create a branch
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/commits/{ref}:
    get:
      operationId: listCommits
      summary: Commit history for a ref
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: ref
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/diff:
    get:
      operationId: getDiff
      summary: Unified diff between two refs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/tags:
    get:
      operationId: listTags
      summary: Tags
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/tree/{ref}/{path}:
    get:
      operationId: getTree
      summary: Tree listing at a path
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: ref
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/user:
    get:
      operationId: getCurrentUser
      summary: The authenticated user
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/2fa/setup:
    post:
      operationId: setup2FA
      summary: Begin TOTP enrolment
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/2fa/verify:
    post:
      operationId: verify2FA
      summary: Confirm TOTP enrolment
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/ssh-keys:
    get:
      operationId: listSSHKeys
      summary: List SSH public keys
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: addSSHKey
      summary: Add an SSH public key
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/ssh-keys/{id}:
    delete:
      operationId: deleteSSHKey
      summary: Remove an SSH public key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/tokens:
    get:
      operationId: listTokens
      summary: List personal access tokens
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createToken
      summary: Create a personal access token
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/user/tokens/{id}:
    delete:
      operationId: deleteToken
      summary: Revoke a personal access token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"